NO-TICKET: execution-engine: clean up history modules
