+++ execution-engine/storage/src/history/trie/tests.rs
@@ -1,0 +1,60 @@
+mod pointer_block {
+    use crate::history::trie::*;
+
+    #[test]
+    fn assignment_and_indexing() {
+        let test_hash = Blake2bHash::new(b"TrieTrieAgain");
+        let leaf_pointer = Some(Pointer::LeafPointer(test_hash));
+        let mut pointer_block = PointerBlock::new();
+        pointer_block[0] = leaf_pointer;
+        pointer_block[RADIX - 1] = leaf_pointer;
+        assert_eq!(leaf_pointer, pointer_block[0]);
+        assert_eq!(leaf_pointer, pointer_block[RADIX - 1]);
+        assert_eq!(None, pointer_block[1]);
+        assert_eq!(None, pointer_block[RADIX - 2]);
+    }
+
+    #[test]
+    #[should_panic]
+    fn assignment_off_end() {
+        let test_hash = Blake2bHash::new(b"TrieTrieAgain");
+        let leaf_pointer = Some(Pointer::LeafPointer(test_hash));
+        let mut pointer_block = PointerBlock::new();
+        pointer_block[RADIX] = leaf_pointer;
+    }
+
+    #[test]
+    #[should_panic]
+    fn indexing_off_end() {
+        let pointer_block = PointerBlock::new();
+        let _val = pointer_block[RADIX];
+    }
+}
+
+mod proptests {
+    use history::trie::gens::*;
+    use proptest::prelude::proptest;
+    use shared::test_utils::test_serialization_roundtrip;
+
+    proptest! {
+        #[test]
+        fn roundtrip_blake2b_hash(hash in blake2b_hash_arb()) {
+            assert!(test_serialization_roundtrip(&hash));
+        }
+
+        #[test]
+        fn roundtrip_trie_pointer(pointer in trie_pointer_arb()) {
+            assert!(test_serialization_roundtrip(&pointer));
+        }
+
+        #[test]
+        fn roundtrip_trie_pointer_block(pointer_block in trie_pointer_block_arb()) {
+            assert!(test_serialization_roundtrip(&pointer_block));
+        }
+
+        #[test]
+        fn roundtrip_trie(trie in trie_arb()) {
+            assert!(test_serialization_roundtrip(&trie));
+        }
+    }
+}
